requirements.txt
PyQt6>=6.6.0
openai>=1.12.0
google-generativeai>=0.4.0
yt-dlp>=2024.3.1
Pillow>=10.0.0
